+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - New `cassandra` output.
 - Field `allowed_verbs` added to the `http_server` input and output.
 - New bloblang function `now`, and method `parse_timestamp`.
+- New bloblang methods `floor` and `round`.
 - The bloblang method `format_timestamp` now supports strings in ISO 8601 format as well as unix epochs with decimal precision up to nanoseconds.
 
 ## Changed